+++ mvc/src/main/java/nextstep/mvc/adaptor/HandlerExecutionAdapter.java
@@ -2,7 +2,7 @@ package nextstep.mvc.adaptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.mapper.tobe.HandlerExecution;
+import nextstep.mvc.handler.tobe.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
 
 public class HandlerExecutionAdapter implements HandlerAdapter {
