+++ mvc/src/main/java/nextstep/mvc/handler/tobe/HandlerMappings.java
@@ -1,9 +1,10 @@
-package nextstep.mvc.mapper.tobe;
+package nextstep.mvc.handler.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import nextstep.mvc.exception.NotFoundException;
 
 public class HandlerMappings {
     private final List<HandlerMapping> handlerMappings = new ArrayList<>();
@@ -21,6 +22,6 @@ public class HandlerMappings {
                 .map(handlerMapping -> handlerMapping.getHandler(request))
                 .filter(Objects::nonNull)
                 .findFirst()
-                .orElseThrow();
+                .orElseThrow(NotFoundException::new);
     }
 }
